Use Evaluation queue in UCB.

+++ players/Evaluator.java
@@ -5,7 +5,7 @@ import gamer.GameState;
 
 interface Evaluator<G extends Game> extends Cloneable {
   // This method must not modify the state.
-  int evaluate(GameState<G> state);
+  double evaluate(GameState<G> state);
 
   public Evaluator<G> clone();
 }
